# ops-ui/openapi/codex-api.openapi.json
{
  "components": {
    "schemas": {
      "AccountResponse": {
        "title": "AccountResponse"
      },
      "ApiKeyLoginRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "apiKey": "sk-proj-example"
          }
        ],
        "properties": {
          "apiKey": {
            "description": "OpenAI API key used for Codex authentication.",
            "examples": [
              "sk-proj-example"
            ],
            "format": "password",
            "minLength": 1,
            "title": "Apikey",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "apiKey"
        ],
        "title": "ApiKeyLoginRequest",
        "type": "object"
      },
...
      "ChannelCreateRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "handle": "@archive-live",
            "name": "Archive Live",
            "youtubeChannelId": "UC_x5XG1OV2P6uZZ5FSM9Ttw"
          }
        ],
        "properties": {
          "handle": {
            "description": "Public channel handle or short identifier.",
            "examples": [
              "@archive-live"
            ],
            "maxLength": 255,
            "minLength": 1,
            "title": "Handle",
            "type": "string"
          },
          "name": {
            "description": "Display name of the channel.",
            "examples": [
              "Archive Live"
            ],
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "youtubeChannelId": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional YouTube channel ID, not the handle.",
            "examples": [
              "UC_x5XG1OV2P6uZZ5FSM9Ttw"
            ],
            "title": "Youtubechannelid"
          }
        },
        "required": [
          "handle",
          "name"
        ],
        "title": "ChannelCreateRequest",
        "type": "object"
      },
      "ChannelResponse": {
        "properties": {
          "handle": {
            "title": "Handle",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sourceApiCallId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceapicallid"
          },
          "sourceJobId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcejobid"
          },
          "streamerId": {
            "title": "Streamerid",
            "type": "integer"
          },
          "uploadsPlaylistId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadsplaylistid"
          },
          "youtubeChannelId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtubechannelid"
          }
        },
        "required": [
          "id",
          "streamerId",
          "handle",
          "name",
          "youtubeChannelId",
          "uploadsPlaylistId",
          "sourceApiCallId",
          "sourceJobId"
        ],
        "title": "ChannelResponse",
        "type": "object"
      },
      "ChannelUpdateRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "handle": "@archive-shorts",
            "name": "Archive Shorts",
            "youtubeChannelId": "UC_x5XG1OV2P6uZZ5FSM9Ttw"
          }
        ],
        "properties": {
          "handle": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New public channel handle or short identifier.",
            "examples": [
              "@archive-shorts"
            ],
            "title": "Handle"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New display name of the channel.",
            "examples": [
              "Archive Shorts"
            ],
            "title": "Name"
          },
          "youtubeChannelId": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New YouTube channel ID. Use null to clear this optional field.",
            "examples": [
              "UC_x5XG1OV2P6uZZ5FSM9Ttw"
            ],
            "title": "Youtubechannelid"
          }
        },
        "title": "ChannelUpdateRequest",
        "type": "object"
      },
...
      "DeleteResponse": {
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginResponse": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "ResolveYouTubeChannelRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "handle": "@GoogleDevelopers"
          }
        ],
        "properties": {
          "handle": {
            "description": "YouTube handle to resolve. The leading @ is optional.",
            "examples": [
              "@GoogleDevelopers"
            ],
            "maxLength": 255,
            "minLength": 1,
            "title": "Handle",
            "type": "string"
          }
        },
        "required": [
          "handle"
        ],
        "title": "ResolveYouTubeChannelRequest",
        "type": "object"
      },
      "ResolveYouTubeChannelResponse": {
        "properties": {
          "channelId": {
            "title": "Channelid",
            "type": "integer"
          },
          "handle": {
            "title": "Handle",
            "type": "string"
          },
          "jobAttemptId": {
            "title": "Jobattemptid",
            "type": "integer"
          },
          "jobId": {
            "title": "Jobid",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sourceApiCallId": {
            "title": "Sourceapicallid",
            "type": "integer"
          },
          "streamerId": {
            "title": "Streamerid",
            "type": "integer"
          },
          "uploadsPlaylistId": {
            "title": "Uploadsplaylistid",
            "type": "string"
          },
          "youtubeChannelId": {
            "title": "Youtubechannelid",
            "type": "string"
          }
        },
        "required": [
          "channelId",
          "streamerId",
          "handle",
          "name",
          "youtubeChannelId",
          "uploadsPlaylistId",
          "sourceApiCallId",
          "jobId",
          "jobAttemptId"
        ],
        "title": "ResolveYouTubeChannelResponse",
        "type": "object"
      },
...
      "RunRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "baseInstructions": "Answer concisely and include file paths when relevant.",
            "developerInstructions": "Do not modify files; only inspect the repository.",
            "model": "gpt-5.5",
            "prompt": "Explain how this project is structured.",
            "reasoningEffort": "medium"
          }
        ],
        "properties": {
          "baseInstructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional base instructions applied before the prompt.",
            "examples": [
              "Answer concisely and include file paths when relevant."
            ],
            "title": "Baseinstructions"
          },
          "developerInstructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional developer instructions for this run.",
            "examples": [
              "Do not modify files; only inspect the repository."
            ],
            "title": "Developerinstructions"
          },
          "model": {
            "anyOf": [
              {
                "enum": [
                  "gpt-5.5",
                  "gpt-5.4",
                  "gpt-5.4-mini"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional model override for this run.",
            "examples": [
              "gpt-5.5"
            ],
            "title": "Model"
          },
          "prompt": {
            "description": "User prompt to send to a Codex thread.",
            "examples": [
              "Explain how this project is structured."
            ],
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "reasoningEffort": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "xhigh"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional reasoning effort override for this run.",
            "examples": [
              "medium"
            ],
            "title": "Reasoningeffort"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "RunRequest",
        "type": "object"
      },
      "RunResponse": {
        "properties": {
          "finalResponse": {
            "title": "Finalresponse",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "reasoningEffort": {
            "title": "Reasoningeffort",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "threadId": {
            "title": "Threadid",
            "type": "string"
          },
          "turnId": {
            "title": "Turnid",
            "type": "string"
          },
          "usage": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Usage"
          }
        },
        "required": [
          "threadId",
          "turnId",
          "status",
          "finalResponse",
          "model",
          "reasoningEffort"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST API wrapper for Codex SDK run/account/logout workflows.",
    "title": "Codex SDK CLI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/channels/{channel_id}": {
      "delete": {
        "operationId": "delete_channel_channels__channel_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Channel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Channel",
        "tags": [
          "channels"
        ]
      },
      "get": {
        "operationId": "get_channel_channels__channel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Channel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Channel",
        "tags": [
          "channels"
        ]
      },
      "patch": {
        "operationId": "update_channel_channels__channel_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Channel Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Channel",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/codex/account": {
      "get": {
        "operationId": "get_codex_account_codex_account_get",
        "parameters": [
          {
            "in": "query",
            "name": "refreshToken",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refreshtoken",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Codex Account",
        "tags": [
          "codex"
        ]
      }
    },
    "/codex/login/api-key": {
      "post": {
        "operationId": "login_codex_with_api_key_codex_login_api_key_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login Codex With Api Key",
        "tags": [
          "codex"
        ]
      }
    },
...
    "/codex/runs": {
      "post": {
        "operationId": "run_codex_prompt_codex_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Codex Prompt",
        "tags": [
          "codex"
        ]
      }
    },
...
    "/streamers/{streamer_id}/channels": {
      "get": {
        "operationId": "list_streamer_channels_streamers__streamer_id__channels_get",
        "parameters": [
          {
            "in": "path",
            "name": "streamer_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Streamer Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChannelResponse"
                  },
                  "title": "Response List Streamer Channels Streamers  Streamer Id  Channels Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Streamer Channels",
        "tags": [
          "channels"
        ]
      },
      "post": {
        "operationId": "create_streamer_channel_streamers__streamer_id__channels_post",
        "parameters": [
          {
            "in": "path",
            "name": "streamer_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Streamer Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Streamer Channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/streamers/{streamer_id}/channels/resolve": {
      "post": {
        "operationId": "resolve_streamer_channel_streamers__streamer_id__channels_resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "streamer_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Streamer Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveYouTubeChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveYouTubeChannelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Streamer Channel",
        "tags": [
          "channels"
        ]
      }
    },
...
  }
}